+++ client/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import App from './app'
-import {init} from './api'
+import {init} from './utils/api'
 
 init()
